requirements.txt
packaging==23.1
Pillow==10.0.0
pip==22.3.1
pytesseract==0.3.10
setuptools==65.5.0